% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncpen_util.R
\name{excluded}
\alias{excluded}
\title{Check whether a pair should be excluded from interactions.}
\usage{
excluded(excluded.pair, a, b)
}
\arguments{
\item{excluded.pair}{a pair.}

\item{a}{first column to be compared.}

\item{b}{second column to be compared.}
}
\value{
TRUE if excluded, FALSE otherwise.
}
\description{
This is internal use only function.
}
