% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{native_cpp_ncpen_fun_}
\alias{native_cpp_ncpen_fun_}
\title{Native point ncpen function.}
\usage{
native_cpp_ncpen_fun_(y_vec, x_mat0, x_std, intc, w_vec0, lam_vec0, r_lam, gam,
  tau, p_max, iter_max, b_eps, k_eps, p_eff, r_eff, family, penalty)
}
\arguments{
\item{y_vec}{.}

\item{x_mat0}{.}

\item{x_std}{.}

\item{intc}{.}

\item{w_vec0}{.}

\item{lam_vec0}{.}

\item{r_lam}{.}

\item{gam}{.}

\item{tau}{.}

\item{p_max}{.}

\item{iter_max}{.}

\item{b_eps}{.}

\item{k_eps}{.}

\item{p_eff}{.}

\item{r_eff}{.}

\item{family}{.}

\item{penalty}{.}
}
\value{
.
}
\description{
This is internal use only function. Manual left blank on purpose.
}
