% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{native_cpp_nr_fun_}
\alias{native_cpp_nr_fun_}
\title{N/A.}
\usage{
native_cpp_nr_fun_(fam, y_vec, x_mat, iter_max, b_eps)
}
\arguments{
\item{fam}{.}

\item{y_vec}{.}

\item{x_mat}{.}

\item{iter_max}{.}

\item{b_eps}{.}
}
\value{
.
}
\description{
This is internal use only function. Manual left blank on purpose.
}
