% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{native_cpp_obj_fun_}
\alias{native_cpp_obj_fun_}
\title{Native object function.}
\usage{
native_cpp_obj_fun_(name, y_vec, x_mat, b_vec, r_eff)
}
\arguments{
\item{name}{.}

\item{y_vec}{.}

\item{x_mat}{.}

\item{b_vec}{.}

\item{r_eff}{.}
}
\value{
.
}
\description{
This is internal use only function.
}
