% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{native_cpp_p_ncpen_fun_}
\alias{native_cpp_p_ncpen_fun_}
\title{Native point ncpen function.}
\usage{
native_cpp_p_ncpen_fun_(y_vec, x_mat, b_vec, w_vec, lam, gam, tau, iter_max,
  b_eps, k_eps, p_eff, r_eff, family, penalty)
}
\arguments{
\item{y_vec}{.}

\item{x_mat}{.}

\item{b_vec}{.}

\item{w_vec}{.}

\item{lam}{.}

\item{gam}{.}

\item{tau}{.}

\item{iter_max}{.}

\item{b_eps}{.}

\item{k_eps}{.}

\item{p_eff}{.}

\item{r_eff}{.}

\item{family}{.}

\item{penalty}{.}
}
\value{
.
}
\description{
This is internal use only function. Manual left blank on purpose.
}
