% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{native_cpp_pen_fun_}
\alias{native_cpp_pen_fun_}
\title{Native Penalty function.}
\usage{
native_cpp_pen_fun_(name, b_vec, lam, gam, tau)
}
\arguments{
\item{name}{.}

\item{b_vec}{.}

\item{lam}{.}

\item{gam}{.}

\item{tau}{.}
}
\value{
.
}
\description{
This is internal use only function. Manual left blank on purpose.
}
