% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{native_cpp_set_dev_mode_}
\alias{native_cpp_set_dev_mode_}
\title{N/A.}
\usage{
native_cpp_set_dev_mode_(dev_mode)
}
\arguments{
\item{dev_mode}{.}
}
\value{
.
}
\description{
This is internal use only function. Manual left blank on purpose.
}
