% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncpen-package.R
\docType{package}
\name{ncpen-package}
\alias{ncpen-package}
\title{ncpen: A package for noncovex penalized estimations.}
\arguments{
\item{y.vec}{numeric vector; samples of dependent variable}

\item{x.mat}{numeric matrix; samples of independent variables}

\item{family}{character; model type; defalut is "gaussian"}
}
\value{
This returns...... If integer overflow
  \url{http://en.wikipedia.org/wiki/Integer_overflow} occurs, the output
  will be NA with a warning. Otherwise it will be a length-one numeric or
  complex vector.

  Zero-length vectors have sum 0 by definition. See
  \url{http://en.wikipedia.org/wiki/Empty_sum} for more details.
}
\description{
\code{ncpen} fits generalized linear models by using penalized maximum likelihood estimation.
It covers Various non-convex penalties such as SCAD and MCP for linear, logistic and poisson regression models.
}
\details{
The ncpen package provides ....
and many
}
\note{
Leave some notes here.
}
\section{ncpen functions}{

The ncpen functions ...
}

\examples{
fam = "lin"
pen = "scad"

a = 3 + 4;
a

}
\references{
Paper 1 by Big Name
}
\seealso{
See this also....
}
\author{
Sunghoon Kwon, Dongshin Kim
}
