% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncpen_cpp_wrap.R
\name{ncpen}
\alias{ncpen}
\title{non-convex penalized estimation for generalized linear models}
\usage{
ncpen(y.vec, x.mat, family = c("gaussian", "binomial", "poisson"),
  penalty = c("scad", "mcp", "lasso", "classo", "sridge", "mbridge", "mlog"),
  lambda = NULL, n.lambda = 100, r.lambda = 0.001, w.lambda = NULL,
  tau = 5, gam = 1e-06, ridge = 1e-06, df.max = 50, proj.min = 100,
  iter.max = 1000, b.eps = 1e-06, k.eps = 1e-04, x.standardize = TRUE,
  intercept = TRUE)
}
\arguments{
\item{y.vec}{numeric vector; samples of dependent variable}

\item{x.mat}{numeric matrix; samples of independent variables}

\item{family}{character; model type; defalut is "gaussian"}
}
\value{
This returns...... If integer overflow
  \url{http://en.wikipedia.org/wiki/Integer_overflow} occurs, the output
  will be NA with a warning. Otherwise it will be a length-one numeric or
  complex vector.

  Zero-length vectors have sum 0 by definition. See
  \url{http://en.wikipedia.org/wiki/Empty_sum} for more details.
}
\description{
\code{ncpen} fits generalized linear models by using penalized maximum likelihood estimation.
It covers Various non-convex penalties such as SCAD and MCP for linear, logistic and poisson regression models.
}
\note{
Leave some notes here.
}
\examples{
fam = "lin"
pen = "scad"

a = 3 + 4;
a

\dontrun{}
}
\references{
Paper 1 by Big Name
}
\seealso{
See this also....
}
\author{
Sunghoon Kwon, Dongshin Kim
}
