% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncpen_cpp_wrap.R
\name{plot.ncpen}
\alias{plot.ncpen}
\title{PLOT.NCPEN}
\usage{
\method{plot}{ncpen}(ncpen.fit, log.scale = FALSE, ...)
}
\arguments{
\item{q.mat}{A numeric matrix....}

\item{l.vec}{A numeric vector....}
}
\value{
This returns...... If integer overflow
  \url{http://en.wikipedia.org/wiki/Integer_overflow} occurs, the output
  will be NA with a warning. Otherwise it will be a length-one numeric or
  complex vector.

  Zero-length vectors have sum 0 by definition. See
  \url{http://en.wikipedia.org/wiki/Empty_sum} for more details.
}
\description{
\code{qlasso.fun} returns ......

This is a generic function: methods can be defined for it directly
or via the \code{\link{Summary}} group generic. For this to work properly,
the arguments \code{abc} should be unnamed, and dispatch is on the
first argument.
}
\note{
Leave some notes here.
}
\examples{
fam = "lin"
pen = "scad"

a = 3 + 4;
a

\dontrun{
qlasso.fun("a")
}

}
\references{
Paper 1 by Big Name
}
\seealso{
See this also....
}
\author{
Sunghoon Kwon, Dongshin Kim
}
