% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncpen_util.R
\name{power.data}
\alias{power.data}
\title{Power Data}
\usage{
power.data(data, power, tail = "_pow")
}
\arguments{
\item{data}{a \code{\link{data.frame}} or \code{\link{matrix}} object.}

\item{power}{power.}

\item{tail}{tail text for column names for powered data. For example, if a column "sales" is powered by 4 (=\code{power})
and \code{tail} is "_pow", then the output column name becomes "sales_pow4".}
}
\value{
This returns an object of \code{\link{matrix}}.
}
\description{
\code{power.data} power data and return a \code{\link{data.frame}} with column names with tail.
}
\examples{
df = data.frame(a = 1:3, b= 4:6);
power.data(df, 2, ".pow");


}
