% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncpen_util.R
\name{same.base}
\alias{same.base}
\title{Check whether column names are derivation of a same base.}
\usage{
same.base(base.cols, a, b)
}
\arguments{
\item{base.cols}{vector of base column names.}

\item{a}{first column to be compared.}

\item{b}{second coumn to be compared.}
}
\value{
TRUE if same base, FALSE otherwise.
}
\description{
This is internal use only function.
}
