% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{native_cpp_ncpen_fun_}
\alias{native_cpp_ncpen_fun_}
\title{Native ncpen function.}
\usage{
native_cpp_ncpen_fun_(y_vec, x_mat0, w_vec0, lam_vec0, gam, tau, alp, d_max,
  iter_max, b_eps, k_eps, p_eff, family, penalty, loc, ob_vec, div)
}
\arguments{
\item{y_vec}{.}

\item{x_mat0}{.}

\item{w_vec0}{.}

\item{lam_vec0}{.}

\item{gam}{.}

\item{tau}{.}

\item{alp}{.}

\item{d_max}{.}

\item{iter_max}{.}

\item{b_eps}{.}

\item{k_eps}{.}

\item{p_eff}{.}

\item{family}{.}

\item{penalty}{.}

\item{loc}{.}

\item{ob_vec}{.}

\item{div}{.}
}
\value{
.
}
\description{
This is internal use only function. Manual left blank on purpose.
}
