% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncpen_package.R
\docType{package}
\name{ncpen-package}
\alias{ncpen-package}
\title{ncpen: A package for non-convex penalized estimation in generalized linear models}
\description{
This package fits the generalized linear models with various non-convex penalties.
A unified algorithm is implemented in \bold{ncpen} based on the convex concave procedure or difference convex algorithm that can be applied to most of existing non-convex penalties.
The available penalties in the package are
the least absolute shrinkage and selection operator(LASSO),
smoothly clipped absolute deviation (SCAD),
minimax concave penalty (MCP),
truncated \eqn{\ell_1}-penalty (TLP),
clipped LASSO (CLASSO),
sparse bridge (SRIDGE),
modified bridge (MBRIDGE),
and modified log (MLOG) penalties.
}
\details{
Accepts a design matrix \eqn{X} and vector of responses \eqn{y},
and produces the regularization path over a grid of values for the tuning parameter \code{lambda}.
Also provides user-friendly processes for plotting, selecting tuning parameters using cross-validation or generalized information criterion (GIC),
\eqn{\ell_2}-regularization, penalty weights, standardization and intercept.
}
\note{
This project is funded by Julian Virtue Professorship from
Center for Applied Research at Graziadio School of Business and Management at
Pepperdine University.
}
\references{
Kim, D., Kwon, S. and Lee, S. (2017). A unified algorithm for various penalized regression models: \bold{R} Package \bold{ncpen}.
}
\author{
Dongshin Kim, Sunghoon Kwon and Sangin Lee
}
